requirements.txt
asgiref==3.9.1
Django==5.2.6
django-extensions==4.1
djangorestframework==3.16.1
python-decouple==3.8
sqlparse==0.5.3
